# packages/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Auth-Ninja API",
    "version": "1.0.0",
    "description": "Shared authentication contract for @auth-ninja/next and AuthNinja.AspNetCore. Sessions use HttpOnly, Secure, SameSite=Strict cookies. State-changing routes require CSRF when enabled."
  },
...
  "paths": {
...
    "/auth/passkeys/register/begin": {
      "post": {
        "operationId": "passkeyRegisterBegin",
        "tags": ["passkeys"],
        "summary": "Begin passkey registration",
        "description": "Return WebAuthn registration options with server challenge.",
        "security": [{ "sessionCookie": [], "csrfHeader": [] }],
        "responses": {
          "200": {
            "description": "Registration options",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WebAuthnOptionsResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/SessionExpired" },
          "403": { "$ref": "#/components/responses/CsrfInvalid" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/passkeys/register/finish": {
      "post": {
        "operationId": "passkeyRegisterFinish",
        "tags": ["passkeys"],
        "summary": "Complete passkey registration",
        "description": "Verify attestation and bind credential to user. Validates origin and RP ID.",
        "security": [{ "sessionCookie": [], "csrfHeader": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebAuthnRegistrationFinishRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Passkey registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PasskeyCredentialResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/SessionExpired" },
          "403": { "$ref": "#/components/responses/CsrfInvalid" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/passkeys/login/begin": {
      "post": {
        "operationId": "passkeyLoginBegin",
        "tags": ["passkeys"],
        "summary": "Begin passkey authentication",
        "description": "Return WebAuthn authentication options. Email optional for discoverable credentials.",
        "security": [{ "csrfHeader": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PasskeyLoginBeginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication options",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WebAuthnOptionsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": { "$ref": "#/components/responses/CsrfInvalid" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/passkeys/login/finish": {
      "post": {
        "operationId": "passkeyLoginFinish",
        "tags": ["passkeys"],
        "summary": "Complete passkey authentication",
        "description": "Verify assertion and establish session with rotated session ID.",
        "security": [{ "csrfHeader": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebAuthnAuthenticationFinishRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "headers": {
              "Set-Cookie": { "$ref": "#/components/headers/SessionSetCookie" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/InvalidCredentials" },
          "403": { "$ref": "#/components/responses/CsrfInvalid" },
          "423": { "$ref": "#/components/responses/AccountLocked" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/passkeys": {
      "get": {
        "operationId": "listPasskeys",
        "tags": ["passkeys"],
        "summary": "List registered passkeys",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Passkey credentials for current user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PasskeyListResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/SessionExpired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/passkeys/{credentialId}": {
      "delete": {
        "operationId": "deletePasskey",
        "tags": ["passkeys"],
        "summary": "Remove a passkey",
        "security": [{ "sessionCookie": [], "csrfHeader": [] }],
        "parameters": [
          {
            "name": "credentialId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "minLength": 1 }
          }
        ],
        "responses": {
          "204": { "description": "Passkey removed" },
          "401": { "$ref": "#/components/responses/SessionExpired" },
          "403": { "$ref": "#/components/responses/CsrfInvalid" },
          "404": {
            "description": "Credential not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "Unable to remove passkey."
                }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_session",
        "description": "HttpOnly, Secure, SameSite=Strict session cookie"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token from GET /auth/csrf"
      }
    },
    "headers": {
      "SessionSetCookie": {
        "description": "Session cookie with rotated session ID",
        "schema": {
          "type": "string",
          "example": "auth_session=...; Path=/; HttpOnly; Secure; SameSite=Strict"
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "VALIDATION_ERROR",
              "message": "Invalid request."
            }
          }
        }
      },
      "InvalidCredentials": {
        "description": "Generic authentication failure",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "INVALID_CREDENTIALS",
              "message": "Invalid email or password."
            }
          }
        }
      },
      "AccountLocked": {
        "description": "Account temporarily locked",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "ACCOUNT_LOCKED",
              "message": "Too many failed attempts. Try again later."
            }
          }
        }
      },
      "SessionExpired": {
        "description": "Session missing or expired",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "SESSION_EXPIRED",
              "message": "Session expired."
            }
          }
        }
      },
...
      "CsrfInvalid": {
        "description": "Missing or invalid CSRF token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "CSRF_INVALID",
              "message": "Invalid CSRF token."
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "RATE_LIMITED",
              "message": "Too many requests."
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "enum": [
          "INVALID_CREDENTIALS",
          "ACCOUNT_LOCKED",
          "SESSION_EXPIRED",
          "MFA_REQUIRED",
          "MFA_INVALID",
          "CSRF_INVALID",
          "RATE_LIMITED",
          "FORBIDDEN",
          "VALIDATION_ERROR"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message"],
        "additionalProperties": false,
        "properties": {
          "code": { "$ref": "#/components/schemas/ErrorCode" },
          "message": {
            "type": "string",
            "description": "Generic user-facing message; must not reveal account existence"
          }
        }
      },
      "User": {
        "type": "object",
        "required": ["id", "email"],
        "additionalProperties": false,
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "mfaEnabled": {
            "type": "boolean",
            "description": "Whether TOTP two-factor authentication is enabled for this user."
          },
          "passkeysEnabled": {
            "type": "boolean",
            "description": "Whether this user has at least one registered WebAuthn passkey."
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "required": ["authenticated", "user"],
        "additionalProperties": false,
        "properties": {
          "authenticated": { "const": true },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
...
      "WebAuthnOptionsResponse": {
        "type": "object",
        "required": ["options"],
        "additionalProperties": false,
        "properties": {
          "options": {
            "type": "object",
            "description": "PublicKeyCredentialCreationOptions or PublicKeyCredentialRequestOptions JSON"
          }
        }
      },
      "WebAuthnRegistrationFinishRequest": {
        "type": "object",
        "required": ["response"],
        "additionalProperties": false,
        "properties": {
          "response": {
            "type": "object",
            "description": "RegistrationResponseJSON from @simplewebauthn/browser"
          }
        }
      },
      "WebAuthnAuthenticationFinishRequest": {
        "type": "object",
        "required": ["response"],
        "additionalProperties": false,
        "properties": {
          "response": {
            "type": "object",
            "description": "AuthenticationResponseJSON from @simplewebauthn/browser"
          }
        }
      },
      "PasskeyLoginBeginRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": { "type": "string", "format": "email" }
        }
      },
      "PasskeyCredentialResponse": {
        "type": "object",
        "required": ["credentialId", "createdAt"],
        "additionalProperties": false,
        "properties": {
          "credentialId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "nickname": { "type": "string", "maxLength": 64 }
        }
      },
      "PasskeySummary": {
        "type": "object",
        "required": ["credentialId", "createdAt"],
        "additionalProperties": false,
        "properties": {
          "credentialId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "nickname": { "type": "string", "maxLength": 64 }
        }
      },
      "PasskeyListResponse": {
        "type": "object",
        "required": ["passkeys"],
        "additionalProperties": false,
        "properties": {
          "passkeys": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PasskeySummary" }
          }
        }
      }
    }
  }
}